Increase number of workers from 4 to 8
Now that we have a computer with 8 cores, we can use 8 workers at once.

My interpretation of my findings from stress-testing was that the reason we were seeing the computer struggle when all N cores were demanded simultaneously, was NOT due to an inherent problem with running N model runs at once, but due to some managerial level of the application being pummelled (say, the R API).

See https://mrc-ide.myjetbrains.com/youtrack/articles/mrc-A-170/Daedalus-Performance-testing for more details / actual millisecond outputs, to see how this conclusion was reached. In short: staggering the requests by a short period (250ms) noticeably improved throughput, and reduced the rate of 500 errors when doing large (>24) batches.